feat(reports): faithful GECK output, Win-1252 text, runtime list cleanup
- Decode all game-authored text as Windows-1252 instead of ASCII/UTF-8
  (EsmStringUtils.DecodeGameText/IsPrintableGameText, plumbed through
  EsmParser, ParsedSubrecord, EsmEditorIdStringReader, RuntimeMemoryContext,
  RuntimeBufferStringExtractor) so accented characters survive round-trip.
- Centralize BSSimpleList<T*> walking on RuntimeMemoryContext and pull
  SPEL/ENCH/ALCH effect-item walking into a shared
  RuntimeEffectItemListReader; recipe and magic readers reuse the walker.
- Add GameStatNormalizer with domain guards for armor DT/DR and effect
  magnitude/area/duration/target/actor-value, applied in EffectRecordHandler
  and runtime item/effect readers; armor reader now disambiguates the
  scaled-uint16 DT layout used by FNV-era runtime builds.
- Fix EFIT schema (Magnitude is float) and CNAM creature schema
  (Health=Int32, AttackDamage=Int16); fix ALCH effect-list offset (64+_s)
  and CONT flags offset (152+_s).
- AMMO records now carry ProjectileFormIds list with mixed-endian DAT2
  parsing and weapon reverse-mapping; multiple projectiles surface in
  GECK item reports.
- Recipe parser flushes RCIL/RCOD components via RCQY count subrecords;
  KEYM parser reads model/value/weight from EDID/FULL/MODL/DATA with
  runtime overlay merging.
- Cross-dump aggregator builds shared KeyLockedDoor and ContainerPlacement
  indexes once, propagates them to per-record reports; ReportValidate
  uses the same builder.
- GeckWorldWriter resolves linked-door destinations against
  PlacedReferenceLocation so source doors expose destination cell, door,
  and label inline. Container reports gain a Placements section.
- Dialogue topic display names backfill from prompt/response text when
  FULL looks editor-id-only; comparison JSON exposes searchText so the
  dialogue table can search inside response bodies.
- Comparison HTML JS equalizes visible build-column widths and refines
  composite key heuristics (Placed, PerkEntry, Effect/Target, named items).
- Refresh runtime-world-cell-probe baselines and DMP test fixtures;
  add ComparisonHtmlRegressionTests, RecordParserHandler/RuntimeParity
  tests, and shared Logger test collection.

+++ src/FalloutXbox360Utils/Core/Formats/Esm/Conversion/Schema/SubrecordEffectSchemas.cs
@@ -57,9 +57,8 @@ internal static class SubrecordEffectSchemas
         };
 
         // EFIT - Effect Item (20 bytes)
-        // PDB: EffectItemData struct — all five fields are T_INT4
         schemas[new SubrecordSchemaRegistry.SchemaKey("EFIT", null, 20)] = new SubrecordSchema(
-            F.Int32("Magnitude"), F.UInt32("Area"), F.UInt32("Duration"),
+            F.Float("Magnitude"), F.UInt32("Area"), F.UInt32("Duration"),
             F.UInt32("Type"), F.Int32("ActorValue"))
         {
             Description = "Effect Item"
